Implemented basic parsing functions. Missing matching src exprs.
We still need to finish the algorithm to match the source expressions.
This is sort of trivial once we figure out the regex.

We also need to hook up all the functions in main function.
Later we need to add exception support. Maybe move to OOP style
if we want to.